Implement hopping (see description of commit)
The extra cookies found in the hopped links are added to a the hopped
key in the json file. Note that this are only the additional cookies the
frontpage did not get. We should refactor the lists in JSON file to keys
by cookie name. This makes it easier to check which are only found in
the hopped.